Unsigned int

+++ server/ToClient.h
@@ -7,16 +7,16 @@ struct ToClient
     std::string str;
     char* plateau;
     bool j1;
-	UINT16 x1;
-    UINT16 y1;
+	unsigned int x1;
+    unsigned int y1;
     bool j2;
-    UINT16 x2;
-    UINT16 y2;
+    unsigned int x2;
+    unsigned int y2;
     bool j3;
-    UINT16 x3;
-    UINT16 y3;
+    unsigned int x3;
+    unsigned int y3;
     bool j4;
-    UINT16 x4;
-    UINT16 y4;
+    unsigned int x4;
+    unsigned int y4;
 };
 #endif //ToClient_h
